test: print a uint64_t timebase with %llu
    test_soc.c:1133: format '%u' expects unsigned int, but argument 2 has type
    'uint64_t' [-Werror=format=]

s5l_timer_t::ticks is uint64_t (soc.h:506). The neighbouring frames count was
already cast; the timebase was not. Only the "warnings as errors" job builds
with -Wformat=2, so the other five jobs and the local build all passed.

Verified by reproducing that job's flags locally -- -O2 -Wall -Wextra -Werror
-Wformat=2 -- which is the check I should have run before pushing 9e8b7ac
rather than after reading the failure.

+++ core/tests/test_soc.c
@@ -1130,8 +1130,9 @@ static void test_skipped_refresh_is_invisible_to_the_guest(void) {
     CHECK(acks > 100u, "only %u interrupts in %u ticks — the devices never "
           "fired, so nothing was compared", acks, TICKS);
     CHECK(fast.timer.ticks > 0u && fast.clcd.frames > 0u,
-          "timebase=%u frames=%llu — no device advanced",
-          fast.timer.ticks, (unsigned long long)fast.clcd.frames);
+          "timebase=%llu frames=%llu — no device advanced",
+          (unsigned long long)fast.timer.ticks,
+          (unsigned long long)fast.clcd.frames);
 
     s5l8900_free(&fast);
     s5l8900_free(&reference);
